Reorganise slack subdirectory
