+++ slack/src/helpers.js
@@ -1,8 +1,8 @@
 require('isomorphic-fetch');
 const aws = require('aws-sdk');
 
-const { sendResponse, buildCoreQuery } = require('../lib/helpers');
-const common = require('../lib/common');
+const { sendResponse, buildCoreQuery } = require('../../lib/helpers');
+const common = require('../../lib/common');
 
 const dynamoDB = new aws.DynamoDB();
