+++ slack/src/index.js
